+++ src/main/groovy/org/spongepowered/asm/gradle/plugins/MixinGradlePlugin.groovy
@@ -1,0 +1,45 @@
+/*
+ * This file is part of MixinGradle, licensed under the MIT License (MIT).
+ *
+ * Copyright (c) SpongePowered <https://www.spongepowered.org>
+ * Copyright (c) contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package org.spongepowered.asm.gradle.plugins
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+/**
+ * This is effectively just the entry point for the MixinGradle plugin, most of
+ * the heavy lifting with respect to the AP is handled by {@link MixinExtension}
+ * and here we merely add the mixin extension to the project.
+ */
+public class MixinGradlePlugin implements Plugin<Project> {
+
+    /* (non-Groovydoc)
+     * @see org.gradle.api.Plugin#apply(java.lang.Object)
+     */
+    @Override
+    void apply(Project project) {
+        // create the mixin extension
+        project.extensions.create('mixin', MixinExtension.class, project)
+    }
+}